+++ design-system/stories/Button/ToggleButton.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import ToggleButton from '../../ui/Button/ToggleButton';
+import ToggleButton from '../../ui/buttons/ToggleButton';
 
 type Story = StoryObj<typeof meta>;
